+++ packages/thumbprint-react/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 0.7.0 - 2019-05-09
+
 ### Added
 
 -   [Minor] Create an `Image` component.